Move Room class to its own file
